Refactor: Move user activity controllers to activitiesController.js
Moved the user activity controllers from userActivitiesController.js to activitiesController.js for better organization and maintainability.
