+++ Server/Controllers/activitiesController.js
@@ -17,9 +17,9 @@ const getActivities = asyncHandler(async (req, res) => {
 });
 
 const getActivity = asyncHandler(async (req, res) => {
-    const { activityId } = req.params;
+    const { id } = req.params;
 
-    const activity = await UserActivity.findByPk(activityId);
+    const activity = await UserActivity.findByPk(id);
 
     if (!activity) {
         throw new ApiError('There is no activity with this id.', 404);
@@ -28,4 +28,41 @@ const getActivity = asyncHandler(async (req, res) => {
     res.json({ activity });
 });
 
-module.exports = { getActivities, getActivity };
+const getUserActivities = asyncHandler(async (req, res) => {
+    const { user } = req;
+
+    const { page = 1, limit = 20 } = req.query;
+    const offset = (page - 1) * limit;
+
+    const activities = await UserActivity.findAll({
+        where: { userId: user.id },
+        limit,
+        offset,
+    });
+
+    res.json({ activities });
+});
+
+const getUserActivity = asyncHandler(async (req, res) => {
+    const { user } = req;
+    const { id } = req.params;
+
+    const activity = await UserActivity.findByPk(id);
+
+    if (!activity) {
+        throw new ApiError('There is no activity with this id.', 404);
+    }
+
+    if (activity.userId != user.id) {
+        throw new ApiError('This activity belongs to another user so you cannot access it.', 403);
+    }
+
+    res.json({ activity });
+})
+
+module.exports = {
+    getActivities,
+    getActivity,
+    getUserActivities,
+    getUserActivity,
+};
