+++ Server/Routes/activitiesRoutes.js
@@ -6,6 +6,6 @@ const router = require("express").Router();
 // I don't think this resourse can be public so i limited it for admin.
 
 router.route('/').get(verifyRole('admin'), getActivities);
-router.route('/:activityId').get(verifyRole('admin'), getActivity);
+router.route('/:id').get(verifyRole('admin'), getActivity);
 
 module.exports = router;
