+++ Server/Routes/userRoutes.js
@@ -2,7 +2,7 @@ const { getAllUsers } = require("../Controllers/userController");
 const { protect } = require('../Controllers/authController');
 const { userPreferncesValidator } = require('../utils/validators/prefernecesValidatior');
 const { getpreferneces, updatePreferences } = require('../Controllers/userPreferenecsController');
-const { getUserActivities, getUserActivity } = require('../Controllers/userActivitiesController');
+const { getUserActivities, getUserActivity } = require('../Controllers/activitiesController');
 
 const router = require("express").Router();
 
@@ -11,6 +11,6 @@ router.get("/preferences", protect, getpreferneces);
 router.put("/preferences", protect, userPreferncesValidator, updatePreferences);
 
 router.route('/activities').get(protect, getUserActivities);
-router.route('/activities/:activityId').get(protect, getUserActivity);
+router.route('/activities/:id').get(protect, getUserActivity);
 
 module.exports = router;
